parametrize validation test
